Refine config flow
